requirements.txt
numpy~=1.21.4
pandas~=1.2.3
streamlit~=0.84.1
Pillow==8.3.2

geopandas~=0.9.0
plotly_express~=0.4.1